Add requirement to send mail to Eiffel Community maillist
New project proposals and proposals to move projects to different
lifecycle stages should be announced on Eiffel Community maillist
in order to increase the visibility.

Once the project creation or project lifecycle process is concluded,
TC chair sends an email to Eiffel Community maillist and announce
the new project or the lifecycle stage to broader Eiffel Community.