Update new repo creation howto be inline with project lifecycle process
New repo creation howto document is outdated and does not contain
information about when the described process could be followed.

Based on the work Eiffel BTC did, the creation of new projects are
done following the process described in project lifecycle document
so this commit adds a note together with the links to corresponding
documentation in howto document.