add-new-repository.md: Use repo template as a template (#113)
We have a template repository, eiffel-repository-template, but
previously it hasn't been marked as a template repo in its settings,
hence the "Use this template" button has been unavailable and copying
of the repo contents has been entirely manual. This setting has now
been enabled and the repo creation documentation should be updated
to reflect this.