Add new document for the 2023 TC meetings
The document has been prepopulated with the preliminary agenda of this
year's first TC meeting on Jan 11.

Last year we reached the size limit of HackMD documents and had to split
the document in two. This isn't being done upfront for 2023 since we've
changed how actions are tracked (we now primarily use GitHub issues) so
it's unlikely we'll hit the limit again.